+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   > support UUID type
   > support JSONB type, use "?::jsonb" instead of "?" for jsonb query
   > support gcloud IAM auth
+* log-exporter: strict schema definition
 
 ### 9.2.1 (4/24/2025 - 5/19/2025) !!! only support java 24